Refactoring types for redux-http module
