+++ src/index.ts
@@ -1,5 +1,5 @@
-export * from '@lib/redux-http';
-import { default as InternalBeccaccino } from '@lib/Beccaccino';
+export * from './redux-http';
+import { default as InternalBeccaccino } from './Beccaccino';
 
 export const Beccaccino = {
   configure: InternalBeccaccino.configure,
